requirements.txt
requests
python-dotenv
pyvmomi==8.0.2